Add Room support

+++ as_user_rating.go
@@ -1,7 +1,7 @@
 package scraper
 
 type UserRating struct {
-	Value               float64 `json:"value"`
+	Value               float32 `json:"value"`
 	RatingCount         int     `json:"ratingCount"`
 	RatingCountList     []int   `json:"ratingCountList"`
 	AriaLabelForRatings string  `json:"ariaLabelForRatings"`
